requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
elastic-transport==9.1.0
elasticsearch==9.1.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pdfminer.six==20250506
pycparser==2.23
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
